+++ twrpdtgen/__main__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python3
+
+from twrpdtgen.twrpdtgen import main
+
+if __name__ == '__main__':
+    main()
